[cargo-guppy] move core logic to a new file
Move `Kind` and `FilterOptions` over -- upcoming commits will add more to this file.

Also hang the `make_resolver` function off of `FilterOptions`.
